move roi_dict fixture to conftest.py
