+++ tests/modules/segmentation/filter/test_roi_filter.py
@@ -3,9 +3,6 @@ import pytest
 import numpy as np
 import pathlib
 
-from ophys_etl.modules.decrosstalk.ophys_plane import (
-    OphysROI)
-
 from ophys_etl.modules.segmentation.filter.roi_filter import (
     ROIBaseFilter,
     ROIAreaFilter,
@@ -13,31 +10,6 @@ from ophys_etl.modules.segmentation.filter.roi_filter import (
 
 from ophys_etl.modules.segmentation.filter.schemas import (
     AreaFilterSchema)
-
-
-@pytest.fixture
-def roi_dict():
-
-    rng = np.random.default_rng(66523)
-    roi_dict = dict()
-
-    for roi_id in range(1, 7, 1):
-        x0 = rng.integers(0, 100)
-        y0 = rng.integers(0, 100)
-        width = rng.integers(10, 11)
-        height = rng.integers(10, 15)
-        mask = np.zeros(width*height).astype(bool)
-        dexes = np.arange(len(mask), dtype=int)
-        chosen = rng.choice(dexes, size=2*roi_id, replace=False)
-        mask[chosen] = True
-        roi = OphysROI(x0=int(x0), width=int(width),
-                       y0=int(y0), height=int(height),
-                       mask_matrix=mask.reshape((height, width)),
-                       roi_id=roi_id,
-                       valid_roi=True)
-        assert roi.area == 2*roi_id
-        roi_dict[roi_id] = roi
-    return roi_dict
 
 
 def test_base_roi_filter():
